Wired 0.7.1 porting — "launchable" milestone (runtime crash/assert fixes) ==========================================================================
Follow-up to Phase 1: the binary compiled, linked and launched, but crashed
or asserted its way through startup (Debug-report dialogs, wx3 sizer/menu/
format-string asserts, wizard logic bugs). All found by launching from the
build tree and iterating on what the user reported; the splash-dangling
crashes were pinned from core dumps captured with `run_wired.sh --gdb`.

Changes
-------

wired/src/save/WiredDocument.cpp
  - Startup crash (SaveCenter): the WiredDocument base ctor called
    saveCenter->RegisterId(id) while the global `saveCenter` was still NULL
    (`saveCenter = new SaveCenter()` not yet assigned). The `if(this)` guard
    in RegisterId was UB-reliance: modern gcc removes it under -Os
    (-fdelete-null-pointer-checks) -> guaranteed SIGSEGV at startup. Guarded
    with `if (saveCenter != NULL)`.

wired/src/gui/Wizard.cpp
  - wx3 sizer asserts: dropped the invalid wxALIGN_RIGHT on the horizontal
    nameSizer/pathSizer (71,73,75,77), the redundant .Center() on
    .Expand() (72,76,87,88), and .Bottom() on the vertical mainSizer (89).
  - OnOkClick tested `!lc_Recent`, but lc_Recent is always created in the
    ctor -> OK always demanded a recent-session selection ("No session
    selected"). Now loads a selected recent session, else treats the typed
    name/path as a new project.
  - NewProject called dir.SetCwd() before checking the dir existed -> wx's
    spurious "Could not set current working directory" dialog on every new
    project. Now DirExists() first, create if needed, then SetCwd(); always
    sets chosenDir to the absolute full path.

wired/src/gui/SettingWindow.cpp
  - Same wx3 sizer asserts: removed wxALIGN_RIGHT / .Right() from the
    horizontal horiSizer (97,98) and wxALIGN_BOTTOM from the vertical
    mainSizer (111).

wired/src/gui/MediaLibrary.cpp
  - Old-style Add() carried wxALIGN_TOP|wxALIGN_BOTTOM|wxALL|wxEXPAND on a
    vertical sizer (177): dropped the two contradictory alignments (moot
    next to wxEXPAND), kept wxALL|wxEXPAND.

wired/src/engine/Settings.cpp
  - Exe-relative data/conf fallback: when /usr/local/share/wired and
    /usr/local/etc/wired don't exist, fall back to `data/` + `conf/` next to
    the executable (wxStandardPaths::GetExecutablePath(), i.e. the build-tree
    wired/src/data and wired/src/conf) instead of ""/"./". This makes the ~54
    ABSOLUTE image macros (L"/ihm/...") and the splash resolve from the build
    tree -- a CWD-relative trick can never fix those. Old behavior kept if
    even the exe-relative dirs are missing.

wired/src/gui/MainApp.cpp  (ShowSplash)
  - SIGSEGV in ShowSplash() (either call) was a DANGLING splash pointer:
    wxSplashScreen installs a *global* event filter
    (src/generic/splash.cpp FilterEvent) that closes AND DELETES itself on
    any key/mouse event anywhere in the app -- typing/clicking in the Wizard
    (or a 120s wxSPLASH_TIMEOUT) destroys it, and the static wxSplashScreen*
    dangles. Root cause pinned from the core (`mov (%rdi),%rax` on a freed
    vtable). Replaced wxSplashScreen with a plain wxFrame + wxStaticBitmap
    that nothing else can destroy; ShowSplash(false) Hide()+Destroy()+NULL.

wired/src/gui/MainWindow.h
  - PLUG_MENU_INDEX_START was 50000; wx3 asserts on menu item IDs >= 0x7fff
    outside the wxID_AUTO range (menucmn.cpp "invalid itemid value"), so
    every plugin menu item asserted at startup. Set to 1000 -- above the
    MainWin_* enum (~95), below wxID_LOWEST (4999).

wired/src/libs/WiredWidgets/src/FaderCtrl.cpp
  - s.Printf("%d", lrintf(GetValue())): lrintf() returns long, and wx3's
    wxString::Format type-checks %d as int (strvararg.h "format specifier
    doesn't match argument type") -> assert on every fader repaint. Cast
    (int)lrintf(...) (lines 114,155,248).

wired/src/gui/Ruler.cpp, plugins/loop_sampler/RulerLoop.cpp
  - Same %d-with-long assert on every ruler repaint (dialog storm + "stuck"
    once the main window was up). %d -> %ld (Ruler: 1+(long)(m/SigNumerator),
    RulerLoop: (long)(m/Beats)).

Workflow (untracked, never committed)
-------------------------------------
- run_wired.sh -- launch the GUI from the build tree without make install:
  LD_LIBRARY_PATH over every src/**/.libs dir + the Settings exe-relative
  fallback above. Runs under gdb BY DEFAULT so any crash auto-writes a full
  backtrace to /tmp/wired_gdb.log and a core to /tmp/wired_gdb.core;
  `--direct` bypasses gdb. Added to rsync_me.sh's git-add exclusion list.
- AGENTS.md -- "Graphical testing (run_wired.sh)" section documents the
  edit -> rebuild -> user-launches-and-reports workflow and every fix above.

Verification
------------
- Full tree `make` exit 0.
- `./run_wired.sh` on b550: splash -> wizard (create project) -> main
  window, no crash, no assert dialogs; 8 internal plugins load and their
  images resolve from the build tree. "Launchable" checklist items 3 & 5
  (start; plugins/images) are met. Item 4 (audio in/out) still needs a
  machine with a working sound card.

+++ wired/src/libs/WiredWidgets/src/FaderCtrl.cpp
@@ -111,7 +111,7 @@ void		FaderCtrl::OnLeftDown(wxMouseEvent& event)
       Label->Show(true);
       wxString s;
       if (IsInteger)
-	s.Printf(wxT("%d"), lrintf(GetValue()));
+	s.Printf(wxT("%d"), (int)lrintf(GetValue()));
       else
 	s.Printf(wxT("%.2f"), GetValue());
       Label->SetLabel(s);
@@ -152,7 +152,7 @@ void		FaderCtrl::OnMouseEvent(wxMouseEvent &event)
     {
       wxString s;
       if (IsInteger)
-	s.Printf(wxT("%d"), lrintf(GetValue()));
+	s.Printf(wxT("%d"), (int)lrintf(GetValue()));
       else
 	s.Printf(wxT("%.2f"), GetValue());
       Label->SetLabel(s);
@@ -245,7 +245,7 @@ void		FaderCtrl::OnKeyDown(wxKeyEvent& event)
       Label->Show(true);
       wxString s;
       if (IsInteger)
-	s.Printf(wxT("%d"), lrintf(GetValue()));
+	s.Printf(wxT("%d"), (int)lrintf(GetValue()));
       else
 	s.Printf(wxT("%.2f"), GetValue());
       Label->SetLabel(s);
